# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MoneyMind AI - Personal Finance Intelligence API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [expense, income, any]
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    put:
      operationId: updateCategory
      tags: [categories]
      summary: Update a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryUpdate"
      responses:
        "200":
          description: Updated category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Budgets ───────────────────────────────────────────────────────────────
  /budgets:
    get:
      operationId: listBudgets
      tags: [budgets]
      summary: List all budgets
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [daily, weekly, monthly, yearly]
      responses:
        "200":
          description: List of budgets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Budget"
    post:
      operationId: createBudget
      tags: [budgets]
      summary: Create a budget
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BudgetInput"
      responses:
        "201":
          description: Created budget
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Budget"
  /budgets/{id}:
    put:
      operationId: updateBudget
      tags: [budgets]
      summary: Update a budget
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BudgetUpdate"
      responses:
        "200":
          description: Updated budget
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Budget"
    delete:
      operationId: deleteBudget
      tags: [budgets]
      summary: Delete a budget
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /budgets/utilization:
    get:
      operationId: getBudgetUtilization
      tags: [budgets]
      summary: Get budget utilization stats
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [daily, weekly, monthly, yearly]
      responses:
        "200":
          description: Budget utilization
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BudgetUtilization"

  # ── Goals ─────────────────────────────────────────────────────────────────
...
  /dashboard/recent-transactions:
    get:
      operationId: getRecentTransactions
      tags: [dashboard]
      summary: Get recent transactions for dashboard
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Recent transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"

  # ── Analytics ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Transaction:
      type: object
      required: [id, type, amount, date, category_id]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [expense, income]
        amount:
          type: number
        date:
          type: string
        time:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        category_id:
          type: integer
        category_name:
          type: ["string", "null"]
        category_color:
          type: ["string", "null"]
        category_icon:
          type: ["string", "null"]
        subcategory:
          type: ["string", "null"]
        payment_method:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        tags:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        is_recurring:
          type: boolean
        recurring_frequency:
          type: ["string", "null"]
        mood:
          type: ["string", "null"]
        need_or_want:
          type: ["string", "null"]
        is_business:
          type: boolean
        is_tax_deductible:
          type: boolean
        receipt_url:
          type: ["string", "null"]
        income_source:
          type: ["string", "null"]
        created_at:
          type: string
...
    Category:
      type: object
      required: [id, name, type]
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [expense, income, any]
        icon:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        budget:
          type: ["number", "null"]
        is_default:
          type: boolean
        sort_order:
          type: integer
        created_at:
          type: string
    CategoryInput:
      type: object
      required: [name, type]
      properties:
        name:
          type: string
        type:
          type: string
          enum: [expense, income, any]
        icon:
          type: string
        color:
          type: string
        description:
          type: string
        budget:
          type: number
        sort_order:
          type: integer
    CategoryUpdate:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [expense, income, any]
        icon:
          type: string
        color:
          type: string
        description:
          type: string
        budget:
          type: number
        sort_order:
          type: integer

    # ── Budget ────────────────────────────────────────────────────────────────
    Budget:
      type: object
      required: [id, period, amount]
      properties:
        id:
          type: integer
        period:
          type: string
          enum: [daily, weekly, monthly, yearly]
        amount:
          type: number
        category_id:
          type: ["integer", "null"]
        category_name:
          type: ["string", "null"]
        name:
          type: ["string", "null"]
        start_date:
          type: ["string", "null"]
        end_date:
          type: ["string", "null"]
        created_at:
          type: string
    BudgetInput:
      type: object
      required: [period, amount]
      properties:
        period:
          type: string
          enum: [daily, weekly, monthly, yearly]
        amount:
          type: number
        category_id:
          type: integer
        name:
          type: string
        start_date:
          type: string
        end_date:
          type: string
    BudgetUpdate:
      type: object
      properties:
        period:
          type: string
          enum: [daily, weekly, monthly, yearly]
        amount:
          type: number
        category_id:
          type: integer
        name:
          type: string
        start_date:
          type: string
        end_date:
          type: string
    BudgetUtilization:
      type: object
      required: [budget_id, period, budget_amount, spent_amount, utilization_percent]
      properties:
        budget_id:
          type: integer
        budget_name:
          type: ["string", "null"]
        period:
          type: string
        budget_amount:
          type: number
        spent_amount:
          type: number
        utilization_percent:
          type: number
        category_id:
          type: ["integer", "null"]
        category_name:
          type: ["string", "null"]
        status:
          type: string
          enum: [safe, warning, exceeded]

    # ── Goal ──────────────────────────────────────────────────────────────────
...